requirements.txt
python-telegram-bot[webhooks,job-queue]==20.*
requests